Ocr2/requirements.txt
torch==2.0.1
torchvision==0.15.2
easyocr==1.6.2
opencv-python==4.8.0.76
Pillow==9.5.0
numpy==1.24.3
scikit-learn==1.2.2
language-tool-python==2.7.1
pdf2image==1.16.3
python-docx==0.8.11
flask==2.2.3
pathlib==1.0.1
tqdm==4.65.0
spacy==3.5.0
pandas==2.0.3
